backend/classification/requirements.txt


mlflow==2.19.0
dvc==3.58.0
scikit-learn==1.6.0
pandas==2.2.3
numpy==2.2.1
seaborn==0.13.2
matplotlib==3.10.0
pytest==8.3.4
pyodbc==5.2.0
PyYAML==6.0.2

xgboost==2.1.3
